Release 0.3.6
0.3.6: Hot Fire
- Rebranded the mod as Deep Space Exploration Vessels
- Deprecated the Nuclear Aerospike engine
- Added the WBR-120 Graphene Radiator
- Added the FLM-1800 Multipurpose Storage Tank. Right-click the tank to
store a variety of different resources. Right-click to change the
endcaps when in the
editor. This 1.875m tank fits inside the hollow octo-truss from Near
Future Construction and can be radially attached.
- The Supernova now requires radiators to function properly. You can
disable heat management for the engine by right-clicking the engine and
accessing the debug
window.
- The Supernova now requires 48,000 EC to start the reactor before the
engine can be used. Right-click the part to start the reactor. The EC
startup requirement
can be disabled in the debug menu.
- The Supernova now has a debug menu. Right-click the part to access the
menu. Pressing Debug Reset will remove any overheating, clear out the
heat sink, shut
off the reactor, and reset a few internal variables so the engine
appears as if it just left the VAB/Hangar.

NOTE: Existing spacecraft built prior to 0.3.6 should not be affected by
the changes but just in case, use the debug menu to disable heat
management and the EC
startup requirement.

IMPORTANT NOTE: To install the mod, DELETE the existing
GameData/WildBlueIndustries/NuclearEngine folder before installing DSEV.
Be sure to copy ALL the files
provided into the GameData folder.

FYI: It would be a good idea for new spacecraft built with this version
of the mod to stock up on SpareParts...